+++ e2e-tests/src/it/alphanet/mod.rs
@@ -1,0 +1,1 @@
+mod withdrawal;
