+++ e2e-tests/src/it/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod bindings;
 pub mod devnet;
+pub mod withdrawals;
